requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
timm
numpy
pandas
pillow
requests
transformers>=4.36.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
datasets>=2.16.0
json-repair>=0.1.0
tqdm
wandb
scipy
scikit-learn
